Support up to 2PiB file size

+++ src/main/java/org/warp/filesponge/Metadata.java
@@ -22,4 +22,4 @@ package org.warp.filesponge;
 /**
  * size -1 = unknown size
  */
-public record Metadata(int size) {}
+public record Metadata(long size) {}
